@@ -1,49 +1,61 @@
 {
-  "name": "remix-starter-reactbricks-ui",
+  "name": "remix-starter-v2",
   "private": true,
+  "sideEffects": false,
+  "type": "module",
   "description": "A simple starter to get started quickly with React Bricks and Remix",
   "version": "1.3.7",
   "author": "Matteo Frana <f2@f2.net>",
   "license": "MIT",
   "scripts": {
     "build": "remix build",
-    "dev": "remix dev",
-    "start": "node -r dotenv/config node_modules/.bin/remix-serve build",
+    "dev": "remix dev --manual",
+    "lint": "eslint --ignore-path .gitignore --cache --cache-location ./node_modules/.cache/eslint .",
+    "start": "remix-serve ./build/index.js",
     "typecheck": "tsc"
   },
   "dependencies": {
-    "@headlessui/react": "^1.7.17",
-    "@remix-run/react": "^1.19.1",
-    "@remix-run/serve": "^1.19.1",
-    "@tailwindcss/forms": "^0.5.4",
-    "dayjs": "^1.11.9",
-    "dotenv": "^16.3.1",
+    "@headlessui/react": "^1.7.18",
+    "@remix-run/css-bundle": "^2.7.2",
+    "@remix-run/node": "^2.7.2",
+    "@remix-run/react": "^2.7.2",
+    "@remix-run/serve": "^2.7.2",
+    "@tailwindcss/forms": "^0.5.7",
+    "dayjs": "^1.11.10",
+    "dotenv": "^16.4.2",
     "email-validator": "^2.0.4",
+    "isbot": "^4.1.0",
     "jsonp": "^0.2.1",
     "pigeon-maps": "^0.21.3",
     "prismjs": "^1.29.0",
     "react": "^18.2.0",
-    "react-bricks": "^4.1.0",
+    "react-bricks": "4.3.0-beta.1",
     "react-dom": "^18.2.0",
-    "react-helmet": "^6.1.0",
-    "react-slick": "^0.29.0",
-    "remix": "^1.19.1",
-    "slick-carousel": "^1.8.1"
+    "react-slick": "^0.30.2",
+    "slick-carousel": "^1.8.1",
+    "tailwindcss": "^3.4.1"
   },
   "devDependencies": {
-    "@remix-run/dev": "^1.19.1",
-    "@types/jsonp": "^0.2.1",
-    "@types/prismjs": "^1.26.0",
-    "@types/react": "^18.2.18",
+    "@remix-run/dev": "^2.7.2",
+    "@types/react": "^18.2.20",
     "@types/react-dom": "^18.2.7",
-    "@types/react-helmet": "^6.1.8",
-    "@types/react-slick": "^0.23.10",
-    "postcss": "^8.4.27",
-    "tailwindcss": "^3.3.3",
+    "@types/react-slick": "^0.23.13",
+    "@typescript-eslint/eslint-plugin": "^6.7.4",
+    "eslint": "^8.38.0",
+    "eslint-import-resolver-typescript": "^3.6.1",
+    "eslint-plugin-import": "^2.28.1",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "typescript": "^5.1.6"
   },
   "engines": {
-    "node": ">=14"
+    "node": ">=18.0.0"
   },
-  "sideEffects": false
+  "prettier": {
+    "printWidth": 80,
+    "semi": false,
+    "singleQuote": true,
+    "trailingComma": "es5"
+  }
 }
