@@ -1,9 +1,9 @@
-import { Text, RichText, Image, types } from "react-bricks/frontend"
+import { Text, RichText, Image, types } from 'react-bricks/frontend'
 
 //=============================
 // Local Types
 //=============================
-type Padding = "big" | "small"
+type Padding = 'big' | 'small'
 
 interface HeroUnitProps {
   padding: Padding
@@ -16,37 +16,37 @@ interface HeroUnitProps {
 //=============================
 const HeroUnit: types.Brick<HeroUnitProps> = ({ padding }) => {
   return (
-    <div className='bg-white dark:bg-gray-900'>
+    <div className="bg-white dark:bg-gray-900">
       <div
         className={`max-w-xl mx-auto px-6 ${
-          padding === "big" ? "py-20" : "py-12"
+          padding === 'big' ? 'py-20' : 'py-12'
         }`}
       >
         <div>
           <Image
-            propName='icon'
-            alt='Icon'
+            propName="icon"
+            alt="Icon"
             maxWidth={200}
             aspectRatio={1}
-            imageClassName='w-20 h-20 mb-5 mx-auto'
+            imageClassName="w-20 h-20 mb-5 mx-auto"
           />
           <Text
             renderBlock={(props) => (
-              <h1 className='text-3xl sm:text-4xl text-center font-black text-gray-900 dark:text-white leading-tight mb-3'>
+              <h1 className="text-3xl sm:text-4xl text-center font-black text-gray-900 dark:text-white leading-tight mb-3">
                 {props.children}
               </h1>
             )}
-            placeholder='Type a title...'
-            propName='title'
+            placeholder="Type a title..."
+            propName="title"
           />
           <RichText
             renderBlock={(props) => (
-              <p className='text-xl text-center leading-relaxed text-gray-700 dark:text-gray-100'>
+              <p className="text-xl text-center leading-relaxed text-gray-700 dark:text-gray-100">
                 {props.children}
               </p>
             )}
-            placeholder='Type a text...'
-            propName='text'
+            placeholder="Type a text..."
+            propName="text"
             allowedFeatures={[
               types.RichTextFeatures.Bold,
               types.RichTextFeatures.Italic,
@@ -55,12 +55,12 @@ const HeroUnit: types.Brick<HeroUnitProps> = ({ padding }) => {
               types.RichTextFeatures.Link,
             ]}
             renderCode={(props) => (
-              <code className='text-sm py-1 px-2 bg-gray-200 dark:bg-gray-700 rounded'>
+              <code className="text-sm py-1 px-2 bg-gray-200 dark:bg-gray-700 rounded">
                 {props.children}
               </code>
             )}
             renderLink={(props) => (
-              <a href={props.href} className='text-sky-500 hover:text-sky-600'>
+              <a href={props.href} className="text-sky-500 hover:text-sky-600">
                 {props.children}
               </a>
             )}
@@ -75,24 +75,24 @@ const HeroUnit: types.Brick<HeroUnitProps> = ({ padding }) => {
 // Brick Schema
 //=============================
 HeroUnit.schema = {
-  name: "my-hero-unit",
-  label: "Hero Unit",
-  tags: ["hero unit", "jumbotron"],
+  name: 'my-hero-unit',
+  label: 'Hero Unit',
+  tags: ['hero unit', 'jumbotron'],
   getDefaultProps: () => ({
-    padding: "big",
-    title: "This is a custom Hero Unit",
+    padding: 'big',
+    title: 'This is a custom Hero Unit',
     text: "We are a hi-tech web development company committed to deliver great products on time. We love to understand our customers' needs and exceed expectations.",
   }),
   sideEditProps: [
     {
-      name: "padding",
-      label: "Padding",
+      name: 'padding',
+      label: 'Padding',
       type: types.SideEditPropType.Select,
       selectOptions: {
         display: types.OptionsDisplay.Select,
         options: [
-          { value: "big", label: "Big Padding" },
-          { value: "small", label: "Small Padding" },
+          { value: 'big', label: 'Big Padding' },
+          { value: 'small', label: 'Small Padding' },
         ],
       },
     },
