@@ -1,5 +1,4 @@
-import { useState } from 'react'
-
+import type { LinksFunction, MetaFunction } from '@remix-run/node'
 import {
   Links,
   LiveReload,
@@ -7,22 +6,23 @@ import {
   Outlet,
   Scripts,
   ScrollRestoration,
-  useNavigate,
   useLoaderData,
+  useNavigate,
+  useRouteError,
 } from '@remix-run/react'
-import type { MetaFunction } from '@remix-run/node'
+import { useState } from 'react'
 import { ReactBricks } from 'react-bricks/frontend'
-import config from './react-bricks/config'
 
-import stylesheet from './tailwind.css'
 import ErrorMessage from './components/ErrorMessage'
+import config from './react-bricks/config'
+import stylesheet from './tailwind.css'
 
-export function links() {
-  return [{ rel: 'stylesheet', href: stylesheet }]
-}
+export const links: LinksFunction = () => [
+  ...(stylesheet ? [{ rel: 'stylesheet', href: stylesheet }] : []),
+]
 
 export const meta: MetaFunction = () => {
-  return { title: 'Remix Blog Starter with React Bricks' }
+  return [{ title: 'Remix Blog Starter with React Bricks' }]
 }
 
 export const loader = () => {
@@ -39,7 +39,7 @@ export const loader = () => {
 
 export default function App() {
   const navigate = useNavigate()
-  const { appId, apiKey, environment } = useLoaderData()
+  const { appId, apiKey, environment } = useLoaderData<typeof loader>()
 
   const savedColorMode =
     typeof window === 'undefined' ? '' : localStorage.getItem('color-mode')
@@ -81,7 +81,7 @@ export default function App() {
       <head>
         <script dangerouslySetInnerHTML={{ __html: clientThemeCode }} />
         <meta charSet="utf-8" />
-        <meta name="viewport" content="width=device-width,initial-scale=1" />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
         <Meta />
         <Links />
       </head>
@@ -96,8 +96,9 @@ export default function App() {
     </html>
   )
 }
-export function ErrorBoundary({ error }: { error: Error }) {
-  console.error(error)
+
+export function ErrorBoundary() {
+  const error = useRouteError()
   return (
     <html>
       <head>
@@ -106,7 +107,7 @@ export function ErrorBoundary({ error }: { error: Error }) {
         <Links />
       </head>
       <body>
-        <ErrorMessage error={error} />
+        <ErrorMessage error={error as Error} />
         <Scripts />
       </body>
     </html>
