@@ -3,9 +3,11 @@ import type { MetaFunction } from '@remix-run/node'
 import { Admin, Editor } from 'react-bricks'
 
 export const meta: MetaFunction = () => {
-  return {
-    title: 'React Bricks Editor',
-  }
+  return [
+    {
+      title: 'React Bricks Editor',
+    },
+  ]
 }
 
 const AdminEditor: React.FC = () => {
