@@ -4,9 +4,11 @@ import type { MetaFunction } from '@remix-run/node'
 import { Admin, Playground } from 'react-bricks'
 
 export const meta: MetaFunction = () => {
-  return {
-    title: 'React Bricks Playground',
-  }
+  return [
+    {
+      title: 'React Bricks Playground',
+    },
+  ]
 }
 
 const AdminPlayground: React.FC = () => {
